% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_2rm_2f.R
\name{agreguj_cawi_ucz_2rm_2f}
\alias{agreguj_cawi_ucz_2rm_2f}
\title{Obliczanie wskaznikow z 2. fali 2. rundy monitoringu na poziomie
zagregowanym}
\usage{
agreguj_cawi_ucz_2rm_2f(wskazniki, grupy)
}
\arguments{
\item{wskazniki}{ramka danych z wynikami 2 rundy monitoringu}

\item{grupy}{ramka danych zawierająca definicje podziałów na grupy -
np. zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}}}
}
\value{
data frame
}
\description{
Funkcja oblicza wartości wskaźników na poziomie zagregowanym
na podstawie ramki danych z wynikami ankiety CAWI z drugiej fali drugiej
rundy monitoringu (absolwenci).
}
\seealso{
\code{\link{agreguj_wskazniki}} oraz przekazywane do niej funkcje
używane do obliczania konkretnych wskaźników zagregowanych:
\itemize{
\item{\code{\link{dane_szkoly_2f}},}
\item{\code{\link{liczba_ucz_2f}},}
\item{\code{\link{liczba_kobiet_2f}},}
\item{\code{\link{liczba_kobiet_2f_waga}},}
\item{\code{\link{liczba_abs_2f}},}
\item{\code{\link{liczba_abs_2f_waga}},}
\item{\code{\link{firma_badawcza_2f}},}
\item{\code{\link{formy_2f}},}
\item{\code{\link{zawod_liczebnosc_2f}},}
\item{\code{\link{zawod_liczebnosc_2f_waga}},}
\item{\code{\link{praca_zarobkowa_2f}},}
\item{\code{\link{praca_zarobkowa_2f_waga}},}
\item{\code{\link{praca_zarobkowa_mies_2f}},}
\item{\code{\link{praca_zarobkowa_mies_2f_waga}},}
\item{\code{\link{bezrobocie_2f}},}
\item{\code{\link{bezrobocie_2f_waga}},}
\item{\code{\link{bezrobocie_mies_2f}},}
\item{\code{\link{bezrobocie_mies_2f_waga}},}
\item{\code{\link{neet_2f}},}
\item{\code{\link{neet_2f_waga}},}
\item{\code{\link{neet_mies_2f}},}
\item{\code{\link{neet_mies_2f_waga}},}
\item{\code{\link{status_2020_10}},}
\item{\code{\link{status_2020_10_waga}},}
\item{\code{\link{kontyn_eduk_2f}},}
\item{\code{\link{kontyn_eduk_2f_waga}},}
\item{\code{\link{praca_zarob_2f}},}
\item{\code{\link{praca_zarob_2f_waga}},}
\item{\code{\link{praca_zarob_zawod_2f}},}
\item{\code{\link{praca_zarob_zawod_2f_waga}},}
\item{\code{\link{praca_zarob_branza_2f}},}
\item{\code{\link{praca_zarob_branza_2f_waga}},}
\item{\code{\link{praca_zad_ogolnie_2f}},}
\item{\code{\link{praca_zad_ogolnie_2f_waga}},}
\item{\code{\link{praca_zad_rozwoj_2f}},}
\item{\code{\link{praca_zad_rozwoj_2f_waga}},}
\item{\code{\link{praca_zad_zycie_2f}},}
\item{\code{\link{praca_zad_zycie_2f_waga}},}
\item{\code{\link{praca_zad_zarobki_2f}},}
\item{\code{\link{praca_zad_zarobki_2f_waga}},}
\item{\code{\link{praca_zad_pewnosc_2f}},}
\item{\code{\link{praca_zad_pewnosc_2f_waga}},}
\item{\code{\link{czas_przed_ogolne_2f}},}
\item{\code{\link{czas_przed_ogolne_2f_waga}},}
\item{\code{\link{czas_przedzaw_teo_2f}},}
\item{\code{\link{czas_przedzaw_teo_2f_waga}},}
\item{\code{\link{czas_przedzaw_prak_2f}},}
\item{\code{\link{czas_przedzaw_prak_2f_waga}},}
\item{\code{\link{war_internet_2f}},}
\item{\code{\link{war_internet_2f_waga}},}
\item{\code{\link{war_komp_2f}},}
\item{\code{\link{war_komp_2f_waga}},}
\item{\code{\link{war_miejsce_2f}},}
\item{\code{\link{war_miejsce_2f_waga}},}
\item{\code{\link{war_pomoc_2f}},}
\item{\code{\link{war_pomoc_2f_waga}}}
}
}
