% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{bezrobocie_mies_2f_waga}
\alias{bezrobocie_mies_2f_waga}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
bezrobocie_mies_2f_waga(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja analogiczna do \code{bezrobocie_mies_2f}, ale zwraca
ważone wartości na potrzeby prezentacji danych w grupie odniesienia
(np. województwo).
}
