% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{czas_przedzaw_teo_2f}
\alias{czas_przedzaw_teo_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
czas_przedzaw_teo_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca rozkład odpowiedzi na pytanie \emph{E5.
Jak to, że nauka w okresie pandemii odbywała się zdalnie, wpłynęło na czas
poświęcany przez Panią/Pana na: 2. teoretyczną naukę przedmiotów zawodowych}.
}
