% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{firma_badawcza_2f}
\alias{firma_badawcza_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
firma_badawcza_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
ciąg znaków
}
\description{
Funkcja przechowująca informację o nazwie firmy realizującej
badanie.
}
