% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{formy_2f}
\alias{formy_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
formy_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca listę form gramatycznych różnych słów lub
wyrażeń, które pojawiają się w raporcie w zależności od typu szkoły lub
liczby obserwacji.
}
