% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{neet_2f}
\alias{neet_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
neet_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca odsetek biernych zawodowo wśród absolwentów.
Bierni zawodowo definiowani są jako absolwenci, którzy w  okresie od lipca do
września nie mieli żadnego epizodu nauki, pracy, praktyk ani poszukiwania
pracy. Obliczone na podstawie pytania \emph{PR1}, a konkretnie podunktów:
\emph{uczył(a) się Pan(i) w szkole, na kursie lub na studiach},
\emph{pracował(a) Pan(i) zarobkowo}, \emph{odbywała(a) Pan(i) bezpłatne
praktyki, staż lub wolontariat} oraz \emph{poszukiwał(a) Pan(i) pracy
zarobkowej}
}
