% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{praca_zad_pewnosc_2f}
\alias{praca_zad_pewnosc_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
praca_zad_pewnosc_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik zadowolenie z obecnej pracy biorąc pod
uwagę wszystkie jej aspekty - pytanie \emph{PR7_5: pod względem pewności, że
ją Pan(i) utrzyma}. Podstawą procentowania są tylko pracujący absolwenci
(PR0 = 1), czyli wszystkie kody za wyjątkiem kodu 7 = Nie dotyczy.
\strong{Uwaga:} w kwestionariuszu kod 7 to "trudno powiedzieć", a w zbiorze
ta odowiedź ma kod 8.
Wskaźnik jest rozkładem odpowiedzi:
\itemize{
\item{"Zdecydowanie zadowoleni + Raczej zadowoleni"}{kod 1 + kod 2}
\item{"Ani zadowolony(a) ani niezadowolony(a)"}{kod 3}
\item{"Zdecydowanie niezadowoleni + Raczej niezadowoleni"}{kod 4 + kod 5}
\item{"Trudno powiedzieć"}{kod 8}
}
}
