% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{praca_zarob_branza_2f}
\alias{praca_zarob_branza_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
praca_zarob_branza_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca odsetek absolwentów, któryz nie pracowali w
wyuczonym zawodzie, ale pracowali w branży, do której wyuczony zawód należy
(PR4_1 = 2 & PR4_2 = 1) spośród wszystkich absolwentów, a nie tylko
pracujących.
}
