% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{praca_zarob_branza_2f_waga}
\alias{praca_zarob_branza_2f_waga}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
praca_zarob_branza_2f_waga(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja analogiczna do \code{praca_zarob_branza_2f}, ale zwraca
ważone wartości na potrzeby prezentacji danych w grupie odniesienia (np.
województwo).
}
