% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{praca_zarob_zawod_2f}
\alias{praca_zarob_zawod_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
praca_zarob_zawod_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca odsetek absolwentów pracujących w wyuczonym
zawodzie (pytanie PR4_1 = 1) spośród wszystkich absolwentów, a nie tylko
pracujących.
}
