% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{praca_zarobkowa_mies_2f}
\alias{praca_zarobkowa_mies_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
praca_zarobkowa_mies_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca odsetek absolwentów pracujących zarobkowo w
okresie \strong{lipiec-wrzesień 2020} na podstawie pytania \emph{PR1. Proszę
spojrzeć na poniższy kalendarz i zaznaczyć miesiące, w których w sumie przez
co najmniej dwa tygodnie (bez względu na liczbę godzin w tygodniu) -
pracował(a) Pan(i) zarobkowo} - w podziale na miesiące.
}
