% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{status_2020_10}
\alias{status_2020_10}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
status_2020_10(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca i przechowująca informację o sytuacji
edukacyjno-zawodowej absolwentów w \strong{październiku} 2020 roku:
\itemize{
\item{"tylko pracował"}{PR1_1=1 & PR1_2=2}
\item{"tylko uczył się"}{PR1_1=2 & PR1_2=1}
\item{"uczył się i pracował"}{PR1_1=1 & PR1_2=1}
\item{"nie uczył się ani nie pracował"}{PR1_1=2 & PR1_2=2}
}
}
