% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{status_2020_10_waga}
\alias{status_2020_10_waga}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
status_2020_10_waga(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja analogiczna do \code{status_2020_10}, ale zwraca ważone
wartości na potrzeby prezentacji danych w grupie odniesienia (np.
województwo).
}
