% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{war_internet_2f}
\alias{war_internet_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
war_internet_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik będący rozkładem ocen warunków do pracy
zdalnej na podstawie pytania \emph{E6. Nauka zdalna w okresie epidemii
wymagała odpowiednich warunków. Jakie miał(a) Pan(i) wtedy możliwości
korzystania z...: 1. dobrego połączenia z internetem}
Wskaźnik jest rozkładem odpowiedzi:
\itemize{
\item{"Niewystarczające"}{W ogóle niewystarczające + Raczej
niewystarczające}
\item{"Wystarczające"}{Całkowicie wystarczające + Raczej wystarczające}
\item{"Niepotrzebne do nauki zdalnej"}{Niepotrzebne do nauki zdalnej}
}
}
