% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm_2f.R
\name{zawod_liczebnosc_2f}
\alias{zawod_liczebnosc_2f}
\title{Obliczanie wskaznikow dla 2. fali badania na poziomie zagregowanym}
\usage{
zawod_liczebnosc_2f(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca liczbę absolwentów danego zawodu. Ponadto,
dodana jest informacja o branży, do której należy zawód.
}
